Moved individual components into separate modules
Implemented a slightly different registry pattern
