+++ .gitignore
@@ -91,3 +91,5 @@ ENV/
 
 # PyCharm
 .idea/
+
+.DS_Store
