+++ pycraft/__main__.py
@@ -1,0 +1,4 @@
+
+if __name__ == '__main__':
+    from pycraft.main import main
+    main()
