+++ pycraft/objects/__init__.py
@@ -1,0 +1,3 @@
+import pycraft.objects.block
+import pycraft.objects.item
+import pycraft.objects.enemy
