--- pycraft/objects/default.py
@@ -1,61 +1,0 @@
-TEXTURE_PATH = 'defaults.png'
-FACES = [
-    ( 0, 1, 0),
-    ( 0,-1, 0),
-    (-1, 0, 0),
-    ( 1, 0, 0),
-    ( 0, 0, 1),
-    ( 0, 0,-1),
-]
-
-
-def tex_coord(x, y, n=4):
-    """ Return the bounding vertices of the texture square.
-
-    """
-    m = 1.0 / n
-    dx = x * m
-    dy = y * m
-    return dx, dy, dx + m, dy, dx + m, dy + m, dx, dy + m
-
-
-def tex_coords(top, bottom, side):
-    """ Return a list of the texture squares for the top, bottom and side.
-
-    """
-    top = tex_coord(*top)
-    bottom = tex_coord(*bottom)
-    side = tex_coord(*side)
-    result = []
-    result.extend(top)
-    result.extend(bottom)
-    result.extend(side * 4)
-    return result
-
-GRASS = tex_coords((1, 0), (0, 1), (0, 0))
-SAND = tex_coords((1, 1), (1, 1), (1, 1))
-BRICK = tex_coords((2, 0), (2, 0), (2, 0))
-STONE = tex_coords((2, 1), (2, 1), (2, 1))
-
-
-from .base import Block
-from ..object_registry import register_block
-
-@register_block("grass")
-class Grass(Block):
-    texture = GRASS
-
-@register_block("sand")
-class Sand(Block):
-    texture = SAND
-
-@register_block("brick")
-class Brick(Block):
-    texture = BRICK
-
-@register_block("stone")
-class Stone(Block):
-    texture = STONE
-
-    def destroy(self):
-        pass
