+++ pycraft/objects/enemy.py
@@ -1,0 +1,6 @@
+from pycraft.objects.object import WorldObject, WorldObjectRegistry
+
+objects = WorldObjectRegistry('enemies')
+
+class Enemy(WorldObject):
+    pass
