+++ pycraft/objects/item.py
@@ -1,0 +1,6 @@
+from pycraft.objects.object import WorldObject, WorldObjectRegistry
+
+objects = WorldObjectRegistry('items')
+
+class Item(WorldObject):
+    pass
