+++ pycraft/world.py
@@ -1,0 +1,325 @@
+import time
+from collections import deque
+
+from noise.perlin import SimplexNoise
+from pyglet import image
+from pyglet.gl import *
+from pyglet.graphics import Batch, TextureGroup
+from pyglet.window import mouse
+
+from pycraft.objects.object import WorldObjectRegistry
+from pycraft.util import normalize, sectorize, cube_vertices
+
+simplex_noise2 = SimplexNoise(256).noise2
+
+TEXTURE_PATH = 'pycraft/objects/textures.png'
+FACES = [
+    ( 0, 1, 0),
+    ( 0,-1, 0),
+    (-1, 0, 0),
+    ( 1, 0, 0),
+    ( 0, 0, 1),
+    ( 0, 0,-1),
+]
+
+world_objects = WorldObjectRegistry()
+brick = world_objects.get('blocks.brick')
+grass = world_objects.get('blocks.grass')
+sand = world_objects.get('blocks.sand')
+stone = world_objects.get('blocks.stone')
+
+
+class World:
+
+    def __init__(self):
+        # A Batch is a collection of vertex lists for batched rendering.
+        self.batch = Batch()
+        # A TextureGroup manages an OpenGL texture.
+        self.group = TextureGroup(image.load(TEXTURE_PATH).get_texture())
+        # A mapping from position to the texture of the block at that position.
+        # This defines all the blocks that are currently in the world.
+        self.objects = {}
+        # Same mapping as `world` but only contains blocks that are shown.
+        self.shown = {}
+        # Mapping from position to a pyglet `VertextList` for all shown blocks.
+        self._shown = {}
+        # Which sector the player is currently in.
+        self.sector = None
+        # Mapping from sector to a list of positions inside that sector.
+        self.sectors = {}
+        # Simple function queue implementation. The queue is populated with
+        # _show_block() and _hide_block() calls
+        self.queue = deque()
+        self.init_gl()
+        self._initialize()
+
+    def init_gl(self):
+        """Basic OpenGL configuration."""
+        # Set the color of "clear", i.e. the sky, in rgba.
+        glClearColor(0.5, 0.69, 1.0, 1)
+        # Enable culling (not rendering) of back-facing facets -- facets that aren't
+        # visible to you.
+        glEnable(GL_CULL_FACE)
+        # Set the texture minification/magnification function to GL_NEAREST (nearest
+        # in Manhattan distance) to the specified texture coordinates. GL_NEAREST
+        # "is generally faster than GL_LINEAR, but it can produce textured images
+        # with sharper edges because the transition between texture elements is not
+        # as smooth."
+        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
+        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
+        self.init_gl_fog()
+
+    def init_gl_fog(self):
+        """Configure the OpenGL fog properties."""
+        # Enable fog. Fog "blends a fog color with each rasterized pixel fragment's
+        # post-texturing color."
+        glEnable(GL_FOG)
+        # Set the fog color.
+        glFogfv(GL_FOG_COLOR, (GLfloat * 4)(0.5, 0.69, 1.0, 1))
+        # Say we have no preference between rendering speed and quality.
+        glHint(GL_FOG_HINT, GL_DONT_CARE)
+        # Specify the equation used to compute the blending factor.
+        glFogi(GL_FOG_MODE, GL_LINEAR)
+        # How close and far away fog starts and ends. The closer the start and end,
+        # the denser the fog in the fog range.
+        glFogf(GL_FOG_START, 20.0)
+        glFogf(GL_FOG_END, 60.0)
+
+    def _initialize(self):
+        """Initialize the world by placing all the blocks."""
+        n = 80  # 1/2 width and height of world
+        s = 1  # step size
+        y = 0  # initial y height
+        for x in range(-n, n + 1, s):
+            for z in range(-n, n + 1, s):
+                # create a layer stone an grass everywhere.
+                self.add_block((x, y - 3, z), stone, immediate=False)
+                if x in (-n, n) or z in (-n, n):
+                    # create outer walls.
+                    for dy in range(-2, 3):
+                        self.add_block((x, y + dy, z), stone, immediate=False)
+                else:
+                    y_max = int((simplex_noise2(x/30, z/30) + 1) * 3)
+                    for y_lvl in range(y - 2, y_max):
+                        if y_lvl < (y_max-1):
+                            block = brick
+                        else:
+                            block = grass
+                        self.add_block((x, y_lvl, z), block, immediate=False)
+
+    def hit_test(self, position, vector, max_distance=8):
+        """Line of sight search from current position. If a block is
+        intersected it is returned, along with the block previously in the line
+        of sight. If no block is found, return None, None.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position to check visibility from.
+        vector : tuple of len 3
+            The line of sight vector.
+        max_distance : int
+            How many blocks away to search for a hit.
+        """
+        m = 8
+        x, y, z = position
+        dx, dy, dz = vector
+        previous = None
+        for _ in range(max_distance * m):
+            key = normalize((x, y, z))
+            if key != previous and key in self.objects:
+                return key, previous
+            previous = key
+            x, y, z = x + dx / m, y + dy / m, z + dz / m
+        return None, None
+
+    def exposed(self, position):
+        """Returns False is given `position` is surrounded on all 6 sides by
+        blocks, True otherwise.
+        """
+        x, y, z = position
+        for dx, dy, dz in FACES:
+            if (x + dx, y + dy, z + dz) not in self.objects:
+                return True
+        return False
+
+    def add_block(self, position, texture, immediate=True):
+        """Add a block with the given `texture` and `position` to the world.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to add.
+        texture : list of len 3
+            The coordinates of the texture squares. Use `tex_coords()` to
+            generate.
+        immediate : bool
+            Whether or not to draw the block immediately.
+        """
+        if position in self.objects:
+            self.remove_block(position, immediate)
+        self.objects[position] = texture
+        self.sectors.setdefault(sectorize(position), []).append(position)
+        if immediate:
+            if self.exposed(position):
+                self.show_block(position)
+            self.check_neighbors(position)
+
+    def remove_block(self, position, immediate=True):
+        """Remove the block at the given `position`.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to remove.
+        immediate : bool
+            Whether or not to immediately remove block from canvas.
+        """
+        del self.objects[position]
+        self.sectors[sectorize(position)].remove(position)
+        if immediate:
+            if position in self.shown:
+                self.hide_block(position)
+            self.check_neighbors(position)
+
+    def check_neighbors(self, position):
+        """Check all blocks surrounding `position` and ensure their visual
+        state is current. This means hiding blocks that are not exposed and
+        ensuring that all exposed blocks are shown. Usually used after a block
+        is added or removed.
+        """
+        x, y, z = position
+        for dx, dy, dz in FACES:
+            key = (x + dx, y + dy, z + dz)
+            if key not in self.objects:
+                continue
+            if self.exposed(key):
+                if key not in self.shown:
+                    self.show_block(key)
+            else:
+                if key in self.shown:
+                    self.hide_block(key)
+
+    def show_block(self, position, immediate=True):
+        """Show the block at the given `position`. This method assumes the
+        block has already been added with add_block()
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to show.
+        immediate : bool
+            Whether or not to show the block immediately.
+        """
+        texture = self.objects[position]
+        self.shown[position] = texture
+        if immediate:
+            self._show_block(position, texture)
+        else:
+            self._enqueue(self._show_block, position, texture)
+
+    def _show_block(self, position, block):
+        """Private implementation of the `show_block()` method.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to show.
+        texture : list of len 3
+            The coordinates of the texture squares. Use `tex_coords()` to
+            generate.
+        """
+        x, y, z = position
+        vertex_data = cube_vertices(x, y, z, 0.5)
+        texture_data = block.texture
+        self._shown[position] = self.batch.add(
+            24, GL_QUADS, self.group,
+            ('v3f/static', vertex_data),
+            ('t2f/static', texture_data))
+
+    def hide_block(self, position, immediate=True):
+        """Hide the block at the given `position`. Hiding does not remove the
+        block from the world.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to hide.
+        immediate : bool
+            Whether or not to immediately remove the block from the canvas.
+        """
+        self.shown.pop(position)
+        if immediate:
+            self._hide_block(position)
+        else:
+            self._enqueue(self._hide_block, position)
+
+    def _hide_block(self, position):
+        """Private implementation of the 'hide_block()` method."""
+        self._shown.pop(position).delete()
+
+    def show_sector(self, sector):
+        """Ensure all blocks in the given sector that should be shown are drawn
+        to the canvas.
+        """
+        for position in self.sectors.get(sector, []):
+            if position not in self.shown and self.exposed(position):
+                self.show_block(position, False)
+
+    def hide_sector(self, sector):
+        """Ensure all blocks in the given sector that should be hidden are
+        removed from the canvas.
+        """
+        for position in self.sectors.get(sector, []):
+            if position in self.shown:
+                self.hide_block(position, False)
+
+    def change_sectors(self, before, after):
+        """Move from sector `before` to sector `after`. A sector is a
+        contiguous x, y sub-region of world. Sectors are used to speed up
+        world rendering.
+        """
+        before_set = set()
+        after_set = set()
+        pad = 4
+        for dx in range(-pad, pad + 1):
+            for dy in [0]:  # range(-pad, pad + 1):
+                for dz in range(-pad, pad + 1):
+                    if dx ** 2 + dy ** 2 + dz ** 2 > (pad + 1) ** 2:
+                        continue
+                    if before:
+                        x, y, z = before
+                        before_set.add((x + dx, y + dy, z + dz))
+                    if after:
+                        x, y, z = after
+                        after_set.add((x + dx, y + dy, z + dz))
+        show = after_set - before_set
+        hide = before_set - after_set
+        for sector in show:
+            self.show_sector(sector)
+        for sector in hide:
+            self.hide_sector(sector)
+
+    def _enqueue(self, func, *args):
+        """Add `func` to the internal queue."""
+        self.queue.append((func, args))
+
+    def _dequeue(self):
+        """Pop the top function from the internal queue and call it."""
+        func, args = self.queue.popleft()
+        func(*args)
+
+    def process_queue(self, ticks_per_sec):
+        """Process the entire queue while taking periodic breaks. This allows
+        the game loop to run smoothly. The queue contains calls to
+        _show_block() and _hide_block() so this method should be called if
+        add_block() or remove_block() was called with immediate=False
+        """
+        start = time.clock()
+        while self.queue and time.clock() - start < 1.0 / ticks_per_sec:
+            self._dequeue()
+
+    def process_entire_queue(self):
+        """Process the entire queue with no breaks."""
+        while self.queue:
+            self._dequeue()
