+++ requirements-dev.txt
@@ -1,0 +1,4 @@
+-r requirements.txt
+coverage==4.0.3
+mock==2.0.0
+pytest==2.9.1
